+++ mcp_tools/dns_enum/dnsenum.py
@@ -1,0 +1,33 @@
+# mcp_tools/dns_enum/dnsenum.py
+
+from typing import Dict, Any
+
+def register_dnsenum_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def dnsenum_scan(domain: str, dns_server: str = "", wordlist: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute dnsenum for DNS enumeration with enhanced logging.
+
+        Args:
+            domain: Target domain
+            dns_server: DNS server to use
+            wordlist: Wordlist for brute forcing
+            additional_args: Additional dnsenum arguments
+
+        Returns:
+            DNS enumeration results
+        """
+        data = {
+            "domain": domain,
+            "dns_server": dns_server,
+            "wordlist": wordlist,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting DNSenum: {domain}")
+        result = hexstrike_client.safe_post("api/tools/dnsenum", data)
+        if result.get("success"):
+            logger.info(f"✅ DNSenum completed for {domain}")
+        else:
+            logger.error(f"❌ DNSenum failed for {domain}")
+        return result
