+++ mcp_tools/web_framework/browser_agent.py
@@ -1,0 +1,52 @@
+# mcp_tools/web_framework/browser_agent.py
+
+from typing import Dict, Any, Optional
+
+def register_browser_agent_tool(mcp, hexstrike_client, logger, HexStrikeColors):
+
+    @mcp.tool()
+    def browser_agent_inspect(url: str, headless: bool = True, wait_time: int = 5,
+                             action: str = "navigate", proxy_port: Optional[int] = None, active_tests: bool = False) -> Dict[str, Any]:
+        """
+        AI-powered browser agent for comprehensive web application inspection and security analysis.
+
+        Args:
+            url: Target URL to inspect
+            headless: Run browser in headless mode
+            wait_time: Time to wait after page load
+            action: Action to perform (navigate, screenshot, close, status)
+            proxy_port: Optional proxy port for request interception
+            active_tests: Run lightweight active reflected XSS tests (safe GET-only)
+
+        Returns:
+            Browser inspection results with security analysis
+        """
+        data_payload = {
+            "url": url,
+            "headless": headless,
+            "wait_time": wait_time,
+            "action": action,
+            "proxy_port": proxy_port,
+            "active_tests": active_tests
+        }
+
+        logger.info(f"{HexStrikeColors.CRIMSON}🌐 Starting Browser Agent {action}: {url}{HexStrikeColors.RESET}")
+        result = hexstrike_client.safe_post("api/tools/browser-agent", data_payload)
+
+        if result.get("success"):
+            logger.info(f"{HexStrikeColors.SUCCESS}✅ Browser Agent {action} completed for {url}{HexStrikeColors.RESET}")
+
+            # Enhanced logging for security analysis
+            if action == "navigate" and result.get("result", {}).get("security_analysis"):
+                security_analysis = result["result"]["security_analysis"]
+                issues_count = security_analysis.get("total_issues", 0)
+                security_score = security_analysis.get("security_score", 0)
+
+                if issues_count > 0:
+                    logger.warning(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Security Issues: {issues_count} | Score: {security_score}/100 {HexStrikeColors.RESET}")
+                else:
+                    logger.info(f"{HexStrikeColors.HIGHLIGHT_GREEN} No security issues found | Score: {security_score}/100 {HexStrikeColors.RESET}")
+        else:
+            logger.error(f"{HexStrikeColors.ERROR}❌ Browser Agent {action} failed for {url}{HexStrikeColors.RESET}")
+
+        return result
